Add remote host data to ts_client and ts_bosh stats.
Turn the stat key into a tuple of {Name, Host} so that rollups happen
using both items, instead of just the name.

WIP: there's some places where bjc_undef is used that need to be
investigated, but they all appear to be about parsing the config, so
using just Name may be appropriate here.